Solve Insert a Node at the Tail of a Linked List
They are sending and expecting HEAD with each iteration. For some reason, I was not able to figure out that and was sending only newNode. That was because I configured listedList for practice differently. I was not expecting HEAD.

Note:
NodeA-->NodeB-->null
When we are traversing over the LinkedList, we need to watch for weather the next Node is null or not. If we watch for current Node, then we will loose the link. And it will show only last element inserted in the list.

```java
 static SinglyLinkedListNode insertNodeAtTail(SinglyLinkedListNode head, int data) {
        SinglyLinkedListNode currentNode = head;
        SinglyLinkedListNode newNode = new SinglyLinkedListNode(data);

        if(currentNode == null){
            head = newNode;
            return head;
        }else{
            while(currentNode.next != null){
                currentNode = currentNode.next;
            }
            currentNode.next = newNode;
        }
        return head;
    }
```