add HackerRank LinkedList Insert a node at a specific position in a linked list
It caught my attention for some time.
Had to insert the element at specific location.
I started a counter at 0 and tracked along with traversing the linked list.
Initially, condition was set to `while(count!=position)`

The problem here was: When count and position will be same, I already at the location
where I had to insert the element.

So I initialized my counter to 1.